Mirror new setup for copy-reference.conf osgi hooks from Akka build.
* spray-osgi wasn't using basic spray settings, needed scala 2.10.x
* created a sequence of projects that defines bundled projects for spray
* only pulling reference.conf from internal projects, and creating new reference.conf

This *should* allow the spary-uber-osgi jar to succeed.  however, I'm still
not sure whether or not we need to aggregate reference.conf files for akka itself
in this jar.  If we do, then we'll have to adapt the reference.conf copying task
to inspect the dependencyClasspath and pull reference.conf files from there
as well.